feat: add standalone playlist covers page at /playlists/covers
- Created a new route `/playlists/covers` to display a persistent playlist covers screensaver.
- Enhanced `PlaylistScreensaver.svelte` with a `standalone` prop to disable activity dismissal and enable interactive links.
- Centralized `PlaylistItem` type in `src/lib/types.ts`.
- Centralized `SPOTIFY_PLAYLIST_LINK` constant in `src/lib/config.ts`.
- Refactored `PlaylistCard.svelte` and the main playlists page to use shared resources.
- Verified functionality with Playwright tests.
